+++ src/main/java/Lession6/EmployeeProcessor.java
@@ -13,14 +13,16 @@ public class EmployeeProcessor {
         searchNameEndingInConsonant(stringArrays);
         printNamesWithoutDuplicateChars(stringArrays);
     }
-//Метод вывода отсортированного списка
+
+    //Метод вывода отсортированного списка
     private static void sortStringArrays(String[] stringArrays) {
         System.out.println("\nВывод отсортированный список сотрудников");
         Arrays.sort(stringArrays);
         System.out.println(Arrays.toString(stringArrays));
 
     }
-//Метод который выводит фамилий с инициалами
+
+    //Метод который выводит фамилий с инициалами
     private static void printInitials(String[] stringArrays) {
         System.out.println("\nВывод фамилий с инициалами");
         for (String stringArray : stringArrays) {
@@ -31,7 +33,8 @@ public class EmployeeProcessor {
             System.out.println(surname + " " + firsnameInitial + "." + secondnameinitial + ".");
         }
     }
-//Метод который выводит ФИО чьи фамилии начинаются на гласную
+
+    //Метод который выводит ФИО чьи фамилии начинаются на гласную
     private static void searchVowelSurnames(String[] stringArrays) {
         System.out.println("\nВывод ФИО чьи фамилии начинаются на гласную");
         for (String stringArray : stringArrays) {
@@ -41,7 +44,8 @@ public class EmployeeProcessor {
             }
         }
     }
-//Метод который выводит ФИО имена котрых заканчиваются на согласную
+
+    //Метод который выводит ФИО имена котрых заканчиваются на согласную
     private static void searchNameEndingInConsonant(String[] stringArrays) {
         System.out.println("\nВывод ФИО чьи имена заканчиваются на согласную ");
         for (String stringArray : stringArrays) {
@@ -51,23 +55,24 @@ public class EmployeeProcessor {
         }
     }
 
-//Метод который проверяет начало фамилии на гласную
+    //Метод который проверяет начало фамилии на гласную
     private static boolean checkStartVowel(String string) {
         String vowels = "АЕЁИОУЫЭЮЯаеёиоуыэюя";
         return vowels.indexOf(string.charAt(0)) != -1;
     }
-//Метод который проверяет окончание имени на согласную
+
+    //Метод который проверяет окончание имени на согласную
     private static boolean checkEndConsonantal(String string) {
         String name = string.split(" ")[1];
         char lastChar = name.charAt(name.length() - 1);
         String consonants = "БВГДЖЗЙКЛМНПРСТФХЦЧШЩбвгджзйклмнпрстфхцчшщ";
         return consonants.indexOf(lastChar) != -1;
     }
-//Метод который удаляет дубликаты букв в строках
+
+    //Метод который удаляет дубликаты букв в строках
     private static String removeDuplicateChars(String string) {
         Set<Character> seen = new HashSet<>();
         StringBuilder result = new StringBuilder();
-
         for (char c : string.toCharArray()) {
             char upperC = Character.toUpperCase(c);
             if (!seen.contains(upperC)) {
@@ -77,11 +82,11 @@ public class EmployeeProcessor {
         }
         return result.toString();
     }
-  //Метод который выводит строки без дубликатов букв
+
+    //Метод который выводит строки без дубликатов букв
     private static void printNamesWithoutDuplicateChars(String[] employees) {
         for (String emp : employees) {
             System.out.println(removeDuplicateChars(emp));
         }
     }
-
 }
